Allow ember app to hit Topics controller methods
Since we are using this Rails app as API only. There was an error
preventing the params to be access in the controller. The reason was
that due to the MIME type, the request was throwing a 500 error.

The fix was to register the type MIME `application/vnd.api+json` so that
`content_type` header has the correct type.

Another error preventing was the relationship constrain of `user_id`
unable to be `nil` for `topic.user_id` which was address in the
controller method `create`.

This was a temporary change for the `@topic.user_id = 1` which we will
change as soon as we are able to.

related:
* https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/MIME_types
* https://github.com/rails/rails/issues/28564