Return front end application based on environment
We need to deploy the front end application from the Rails app to avoid
CORS issues. If the app is running on development we use reverse_proxy
to redirect to localhost:4200.

If the production application is running, we use the URI library to make
a request to the S3 endpoint, return the ember app and render it. Thus,
avoiding CORS issues